requirements.txt
pandas
numpy
requests
scikit-learn
streamlit
python-dotenv
matplotlib
seaborn